refactor(cli): keep local workspace hub internals package-local
