+++ packages/cli/package.json
@@ -20,9 +20,11 @@
   },
   "dependencies": {
     "@hachej/boring-agent": "workspace:*",
+    "@hachej/boring-ui-kit": "workspace:*",
     "@hachej/boring-workspace": "workspace:*",
     "@fastify/static": "^8.0.0",
-    "fastify": "^5.4.0"
+    "fastify": "^5.4.0",
+    "lucide-react": "^1.8.0"
   },
   "devDependencies": {
     "@tailwindcss/vite": "^4.0.0",
