+++ packages/cli/src/__tests__/localWorkspaces.test.ts
@@ -2,7 +2,7 @@ import { mkdtemp, rm, writeFile } from "node:fs/promises"
 import { tmpdir } from "node:os"
 import { join } from "node:path"
 import { afterEach, expect, test } from "vitest"
-import { createLocalWorkspaceRegistry } from "../localWorkspaces.js"
+import { createLocalWorkspaceRegistry } from "../server/localWorkspaces.js"
 
 const tempDirs: string[] = []
