+++ packages/cli/src/front/WorkspaceSwitcherControl.tsx
