+++ packages/workspace/src/app/front/index.ts
@@ -10,11 +10,6 @@ export {
   type WorkspaceBootGateProps,
 } from "./WorkspaceBootGate"
 export {
-  WorkspaceSwitcherControl,
-  type WorkspaceSwitcherControlItem,
-  type WorkspaceSwitcherControlProps,
-} from "./WorkspaceSwitcherControl"
-export {
   createLocalStorageSessions,
   useLocalStorageSessions,
   type CreateLocalStorageSessionsOptions,
